+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/adduser/AddUserTests.html
@@ -9,7 +9,6 @@
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>Test Suite</b></td></tr>
 <tr><td><a href="AddUserTest.html">AddUserTest</a></td></tr>
-<tr><td><a href="EditUserPasswordTest.html">EditUserPasswordTest</a></td></tr>
 <tr><td><a href="TearDownUserTest.html">TearDownUserTest</a></td></tr>
 </tbody></table>
 </body>
