--- portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/adduserpassword/AddUserPasswordTest.html
@@ -1,94 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>AddUserPasswordTest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">AddUserPasswordTest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/web/guest/home/</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>link=Control Panel</td>
-	<td>Control Panel</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Control Panel</td>
-	<td>Control Panel</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Users and Organizations</td>
-	<td>Users and Organizations</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_125_keywords']</td>
-	<td>selen01</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//input[@value='Search']</td>
-	<td>Search</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//td[2]/a</td>
-	<td>User Name</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//a[@id='_125_passwordLink']</td>
-	<td>Password</td>
-</tr>
-<tr>
-	<td>assertPartialText</td>
-	<td>//a[@id='_125_passwordLink']</td>
-	<td>Password</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//a[@id='_125_passwordLink']</td>
-	<td>Password</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//input[@id='_125_password1']</td>
-	<td></td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_125_password1']</td>
-	<td>asdf</td>
-</tr>
-<tr>
-	<td>type</td>
-	<td>//input[@id='_125_password2']</td>
-	<td>asdf</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//input[@value='Save']</td>
-	<td>Save</td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-
-</tbody></table>
-</body>
-</html>
