--- portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/adduserpassword/AddUserTest.java
@@ -1,379 +1,0 @@
-/**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portalweb.portal.controlpanel.users.user.adduserpassword;
-
-import com.liferay.portalweb.portal.BaseTestCase;
-import com.liferay.portalweb.portal.util.RuntimeVariables;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class AddUserTest extends BaseTestCase {
-	public void testAddUser() throws Exception {
-		int label = 1;
-
-		while (label >= 1) {
-			switch (label) {
-			case 1:
-				selenium.open("/web/guest/home/");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("link=Control Panel")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.clickAt("link=Control Panel",
-					RuntimeVariables.replace("Control Panel"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				selenium.clickAt("link=Users and Organizations",
-					RuntimeVariables.replace("Users and Organizations"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				selenium.clickAt("link=Add", RuntimeVariables.replace("Add"));
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible(
-									"//div[@class='lfr-component lfr-menu-list']/ul/li/a")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				assertEquals(RuntimeVariables.replace("User"),
-					selenium.getText(
-						"//div[@class='lfr-component lfr-menu-list']/ul/li/a"));
-				selenium.click(RuntimeVariables.replace(
-						"//div[@class='lfr-component lfr-menu-list']/ul/li/a"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				selenium.select("//select[@id='_125_prefixId']",
-					RuntimeVariables.replace("label=Mr."));
-				selenium.type("//input[@id='_125_screenName']",
-					RuntimeVariables.replace("selenium01"));
-				selenium.type("//input[@id='_125_emailAddress']",
-					RuntimeVariables.replace("test01@selenium.com"));
-				selenium.type("//input[@id='_125_firstName']",
-					RuntimeVariables.replace("selen01"));
-				selenium.type("//input[@id='_125_middleName']",
-					RuntimeVariables.replace("lenn"));
-				selenium.type("//input[@id='_125_lastName']",
-					RuntimeVariables.replace("nium01"));
-				selenium.select("//select[@id='_125_suffixId']",
-					RuntimeVariables.replace("label=Phd."));
-				selenium.select("//select[@id='_125_birthdayMonth']",
-					RuntimeVariables.replace("label=April"));
-				selenium.select("//select[@id='_125_birthdayDay']",
-					RuntimeVariables.replace("label=10"));
-				selenium.select("//select[@id='_125_birthdayYear']",
-					RuntimeVariables.replace("label=1986"));
-				selenium.select("//select[@id='_125_male']",
-					RuntimeVariables.replace("label=Male"));
-				selenium.clickAt("//input[@value='Save']",
-					RuntimeVariables.replace("Save"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible(
-									"//div[@class='portlet-msg-success']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				assertEquals(RuntimeVariables.replace(
-						"Your request completed successfully."),
-					selenium.getText("//div[@class='portlet-msg-success']"));
-				assertEquals("selenium01",
-					selenium.getValue("//input[@id='_125_screenName']"));
-				assertEquals("test01@selenium.com",
-					selenium.getValue("//input[@id='_125_emailAddress']"));
-				assertEquals("selen01",
-					selenium.getValue("//input[@id='_125_firstName']"));
-				assertEquals("lenn",
-					selenium.getValue("//input[@id='_125_middleName']"));
-				assertEquals("nium01",
-					selenium.getValue("//input[@id='_125_lastName']"));
-				assertTrue(selenium.isPartialText(
-						"//a[@id='_125_passwordLink']", "Password"));
-				selenium.clickAt("//a[@id='_125_passwordLink']",
-					RuntimeVariables.replace("Password"));
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("//input[@id='_125_password1']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.type("//input[@id='_125_password1']",
-					RuntimeVariables.replace("password"));
-				selenium.type("//input[@id='_125_password2']",
-					RuntimeVariables.replace("password"));
-				selenium.clickAt("//input[@value='Save']",
-					RuntimeVariables.replace("Save"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible(
-									"//div[@class='portlet-msg-success']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				assertEquals(RuntimeVariables.replace(
-						"Your request completed successfully."),
-					selenium.getText("//div[@class='portlet-msg-success']"));
-				selenium.open("/web/guest/home/");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("link=Sign Out")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.clickAt("link=Sign Out",
-					RuntimeVariables.replace("Sign Out"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				selenium.open("/web/guest/home/");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("//input[@id='_58_login']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.type("//input[@id='_58_login']",
-					RuntimeVariables.replace("test01@selenium.com"));
-				selenium.type("//input[@id='_58_password']",
-					RuntimeVariables.replace("password"));
-
-				boolean rememberMeCheckboxChecked1 = selenium.isChecked(
-						"_58_rememberMeCheckbox");
-
-				if (rememberMeCheckboxChecked1) {
-					label = 2;
-
-					continue;
-				}
-
-				selenium.clickAt("//input[@id='_58_rememberMeCheckbox']",
-					RuntimeVariables.replace("Remember Me Checkbox"));
-
-			case 2:
-				selenium.clickAt("//input[@value='Sign In']",
-					RuntimeVariables.replace("Sign In"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-				boolean iAgreeVisible1 = selenium.isElementPresent(
-						"//span/input");
-
-				if (!iAgreeVisible1) {
-					label = 3;
-
-					continue;
-				}
-
-				selenium.clickAt("//input[@value='I Agree']",
-					RuntimeVariables.replace("I Agree"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-			case 3:
-
-				boolean newPasswordVisible1 = selenium.isElementPresent(
-						"//span/input");
-
-				if (!newPasswordVisible1) {
-					label = 4;
-
-					continue;
-				}
-
-				selenium.type("password1", RuntimeVariables.replace("test"));
-				selenium.type("password2", RuntimeVariables.replace("test"));
-				selenium.clickAt("//input[@value='Save']",
-					RuntimeVariables.replace("Save"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-			case 4:
-
-				boolean passwordReminderVisible1 = selenium.isElementPresent(
-						"reminderQueryAnswer");
-
-				if (!passwordReminderVisible1) {
-					label = 5;
-
-					continue;
-				}
-
-				assertEquals(RuntimeVariables.replace(
-						"Please choose a reminder query."),
-					selenium.getText("//form/div[1]"));
-				selenium.type("reminderQueryAnswer",
-					RuntimeVariables.replace("test"));
-				selenium.clickAt("//input[@value='Save']",
-					RuntimeVariables.replace("Save"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-			case 5:
-				selenium.open("/web/guest/home/");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("link=Sign Out")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.clickAt("link=Sign Out",
-					RuntimeVariables.replace("Sign Out"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				selenium.open("/web/guest/home/");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("//input[@id='_58_login']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.type("//input[@id='_58_login']",
-					RuntimeVariables.replace("test@liferay.com"));
-				selenium.type("//input[@id='_58_password']",
-					RuntimeVariables.replace("test"));
-
-				boolean rememberMeCheckboxChecked2 = selenium.isChecked(
-						"_58_rememberMeCheckbox");
-
-				if (rememberMeCheckboxChecked2) {
-					label = 6;
-
-					continue;
-				}
-
-				selenium.clickAt("//input[@id='_58_rememberMeCheckbox']",
-					RuntimeVariables.replace("Remember Me Checkbox"));
-
-			case 6:
-				selenium.clickAt("//input[@value='Sign In']",
-					RuntimeVariables.replace("Sign In"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-			case 100:
-				label = -1;
-			}
-		}
-	}
-}
