--- portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/adduserpassword/TearDownUserTest.html
@@ -1,194 +1,0 @@
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>TearDownUserTest</title>
-</head>
-<body>
-<table cellpadding="1" cellspacing="1" border="1">
-<thead>
-<tr><td rowspan="1" colspan="3">TearDownUserTest</td></tr>
-</thead><tbody>
-<tr>
-	<td>open</td>
-	<td>/web/guest/home/</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForElementPresent</td>
-	<td>link=Control Panel</td>
-	<td>Control Panel</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Control Panel</td>
-	<td>Control Panel</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Users and Organizations</td>
-	<td>Users and Organizations</td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>xPath=(//input[@name='_125_allRowIds'])[2]</td>
-	<td>usersPresent</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${usersPresent}</td>
-	<td>UsersNotPresent</td>
-</tr>
-<tr>
-	<td>assertNotChecked</td>
-	<td>xPath=(//input[@name='_125_allRowIds'])[2]</td>
-	<td>Select All</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>xPath=(//input[@name='_125_allRowIds'])[2]</td>
-	<td>Select All</td>
-</tr>
-<tr>
-	<td>assertChecked</td>
-	<td>xPath=(//input[@name='_125_allRowIds'])[2]</td>
-	<td>Select All</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//input[@value='Deactivate']</td>
-	<td>Deactivate</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to deactivate the selected users?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>UsersNotPresent</td>
-	<td></td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>link=Search All Users</td>
-	<td>Search All Users</td>
-</tr>
-<tr>
-	<td>store</td>
-	<td>eval(selenium.isVisible('link=Advanced »'))</td>
-	<td>advancedVisible</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${advancedVisible}</td>
-	<td>AdvancedNotVisible</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>link=Advanced »</td>
-	<td>Advanced</td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>AdvancedNotVisible</td>
-	<td></td>
-</tr>
-<tr>
-	<td>waitForVisible</td>
-	<td>//select[@id='_125_status']</td>
-	<td>Status</td>
-</tr>
-<tr>
-	<td>select</td>
-	<td>//select[@id='_125_status']</td>
-	<td>Inactive</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//input[@value='Search']</td>
-	<td>Search</td>
-</tr>
-<tr>
-	<td>storeElementPresent</td>
-	<td>//input[@name='_125_allRowIds']</td>
-	<td>usersNotDeleted</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${usersNotDeleted}</td>
-	<td>UsersDeleted</td>
-</tr>
-<tr>
-	<td>assertNotChecked</td>
-	<td>//input[@name='_125_allRowIds']</td>
-	<td>Select All</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>//input[@name='_125_allRowIds']</td>
-	<td>Select All</td>
-</tr>
-<tr>
-	<td>assertChecked</td>
-	<td>//input[@name='_125_allRowIds']</td>
-	<td>Select All</td>
-</tr>
-<tr>
-	<td>clickAtAndWait</td>
-	<td>//input[@value='Delete']</td>
-	<td>Delete</td>
-</tr>
-<tr>
-	<td>assertConfirmation</td>
-	<td>Are you sure you want to permanently delete the selected users?</td>
-	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-success']</td>
-	<td>Your request completed successfully.</td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>UsersDeleted</td>
-	<td></td>
-</tr>
-<tr>
-	<td>assertText</td>
-	<td>//div[@class='portlet-msg-info']</td>
-	<td>No users were found.</td>
-</tr>
-<tr>
-	<td>store</td>
-	<td>eval(selenium.isVisible('link=« Basic'))</td>
-	<td>basicVisible</td>
-</tr>
-<tr>
-	<td>gotoIf</td>
-	<td>!${basicVisible}</td>
-	<td>BasicNotVisible</td>
-</tr>
-<tr>
-	<td>clickAt</td>
-	<td>link=« Basic</td>
-	<td>Basic</td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>BasicNotVisible</td>
-	<td></td>
-</tr>
-
-</tbody></table>
-</body>
-</html>
