--- portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/adduserpassword/TearDownUserTest.java
@@ -1,189 +1,0 @@
-/**
- * Copyright (c) 2000-2012 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portalweb.portal.controlpanel.users.user.adduserpassword;
-
-import com.liferay.portalweb.portal.BaseTestCase;
-import com.liferay.portalweb.portal.util.RuntimeVariables;
-
-/**
- * @author Brian Wing Shun Chan
- */
-public class TearDownUserTest extends BaseTestCase {
-	public void testTearDownUser() throws Exception {
-		int label = 1;
-
-		while (label >= 1) {
-			switch (label) {
-			case 1:
-				selenium.open("/web/guest/home/");
-				loadRequiredJavaScriptModules();
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isElementPresent("link=Control Panel")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.clickAt("link=Control Panel",
-					RuntimeVariables.replace("Control Panel"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				selenium.clickAt("link=Users and Organizations",
-					RuntimeVariables.replace("Users and Organizations"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-				boolean usersPresent = selenium.isElementPresent(
-						"xPath=(//input[@name='_125_allRowIds'])[2]");
-
-				if (!usersPresent) {
-					label = 2;
-
-					continue;
-				}
-
-				assertFalse(selenium.isChecked(
-						"xPath=(//input[@name='_125_allRowIds'])[2]"));
-				selenium.clickAt("xPath=(//input[@name='_125_allRowIds'])[2]",
-					RuntimeVariables.replace("Select All"));
-				assertTrue(selenium.isChecked(
-						"xPath=(//input[@name='_125_allRowIds'])[2]"));
-				selenium.clickAt("//input[@value='Deactivate']",
-					RuntimeVariables.replace("Deactivate"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				assertTrue(selenium.getConfirmation()
-								   .matches("^Are you sure you want to deactivate the selected users[\\s\\S]$"));
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible(
-									"//div[@class='portlet-msg-success']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				assertEquals(RuntimeVariables.replace(
-						"Your request completed successfully."),
-					selenium.getText("//div[@class='portlet-msg-success']"));
-
-			case 2:
-				selenium.clickAt("link=Search All Users",
-					RuntimeVariables.replace("Search All Users"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-				boolean advancedVisible = selenium.isVisible(
-						"link=Advanced \u00bb");
-
-				if (!advancedVisible) {
-					label = 3;
-
-					continue;
-				}
-
-				selenium.clickAt("link=Advanced \u00bb",
-					RuntimeVariables.replace("Advanced"));
-
-			case 3:
-
-				for (int second = 0;; second++) {
-					if (second >= 90) {
-						fail("timeout");
-					}
-
-					try {
-						if (selenium.isVisible("//select[@id='_125_status']")) {
-							break;
-						}
-					}
-					catch (Exception e) {
-					}
-
-					Thread.sleep(1000);
-				}
-
-				selenium.select("//select[@id='_125_status']",
-					RuntimeVariables.replace("Inactive"));
-				selenium.clickAt("//input[@value='Search']",
-					RuntimeVariables.replace("Search"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-
-				boolean usersNotDeleted = selenium.isElementPresent(
-						"//input[@name='_125_allRowIds']");
-
-				if (!usersNotDeleted) {
-					label = 4;
-
-					continue;
-				}
-
-				assertFalse(selenium.isChecked(
-						"//input[@name='_125_allRowIds']"));
-				selenium.clickAt("//input[@name='_125_allRowIds']",
-					RuntimeVariables.replace("Select All"));
-				assertTrue(selenium.isChecked("//input[@name='_125_allRowIds']"));
-				selenium.clickAt("//input[@value='Delete']",
-					RuntimeVariables.replace("Delete"));
-				selenium.waitForPageToLoad("30000");
-				loadRequiredJavaScriptModules();
-				assertTrue(selenium.getConfirmation()
-								   .matches("^Are you sure you want to permanently delete the selected users[\\s\\S]$"));
-				assertEquals(RuntimeVariables.replace(
-						"Your request completed successfully."),
-					selenium.getText("//div[@class='portlet-msg-success']"));
-
-			case 4:
-				assertEquals(RuntimeVariables.replace("No users were found."),
-					selenium.getText("//div[@class='portlet-msg-info']"));
-
-				boolean basicVisible = selenium.isVisible("link=\u00ab Basic");
-
-				if (!basicVisible) {
-					label = 5;
-
-					continue;
-				}
-
-				selenium.clickAt("link=\u00ab Basic",
-					RuntimeVariables.replace("Basic"));
-
-			case 5:
-			case 100:
-				label = -1;
-			}
-		}
-	}
-}
