+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/edituserpassword/EditUser1PasswordTest.html
