+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/edituserpassword/EditUser1PasswordTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.controlpanel.users.user.adduser;
+package com.liferay.portalweb.portal.controlpanel.users.user.edituserpassword;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,8 +20,8 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class EditUser3PasswordTest extends BaseTestCase {
-	public void testEditUser3Password() throws Exception {
+public class EditUser1PasswordTest extends BaseTestCase {
+	public void testEditUser1Password() throws Exception {
 		int label = 1;
 
 		while (label >= 1) {
@@ -93,14 +93,14 @@ public class EditUser3PasswordTest extends BaseTestCase {
 
 			case 2:
 				selenium.type("//input[@name='_125_keywords']",
-					RuntimeVariables.replace("userfn3"));
+					RuntimeVariables.replace("userfn1"));
 				selenium.clickAt("//input[@value='Search']",
 					RuntimeVariables.replace("Search"));
 				selenium.waitForPageToLoad("30000");
 				loadRequiredJavaScriptModules();
-				assertEquals(RuntimeVariables.replace("userfn3"),
+				assertEquals(RuntimeVariables.replace("userfn1"),
 					selenium.getText("//td/a"));
-				selenium.clickAt("//td/a", RuntimeVariables.replace("userfn3"));
+				selenium.clickAt("//td/a", RuntimeVariables.replace("userfn1"));
 				selenium.waitForPageToLoad("30000");
 				loadRequiredJavaScriptModules();
 				assertTrue(selenium.isPartialText(
@@ -140,7 +140,7 @@ public class EditUser3PasswordTest extends BaseTestCase {
 				selenium.waitForPageToLoad("30000");
 				loadRequiredJavaScriptModules();
 				selenium.type("//input[@id='_58_login']",
-					RuntimeVariables.replace("userea3@liferay.com"));
+					RuntimeVariables.replace("userea1@liferay.com"));
 				selenium.type("//input[@id='_58_password']",
 					RuntimeVariables.replace("password"));
