+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/edituserpassword/EditUserPasswordTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.controlpanel.users.user.adduser;
+package com.liferay.portalweb.portal.controlpanel.users.user.edituserpassword;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
