+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/edituserpassword/EditUserPasswordTests.html
@@ -3,14 +3,14 @@
 <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
 <head>
   <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>AddUserPasswordTests</title>
+  <title>Test Suite</title>
 </head>
 <body>
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>AddUserPasswordTests</b></td></tr>
-<tr><td><a href="AddUserTest.html">AddUserTest</a></td></tr>
-<tr><td><a href="AddUserPasswordTest.html">AddUserPasswordTest</a></td></tr>
-<tr><td><a href="TearDownUserTest.html">TearDownUserTest</a></td></tr>
+<tr><td><b>Test Suite</b></td></tr>
+<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUserTest.html">AddUserTest</a></td></tr>
+<tr><td><a href="EditUserPasswordTest.html">EditUserPasswordTest</a></td></tr>
+<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/TearDownUserTest.html">TearDownUserTest</a></td></tr>
 </tbody></table>
 </body>
 </html>
