+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/signin/SignInTests.html
@@ -8,9 +8,12 @@
 <body>
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>Test Suite</b></td></tr>
-<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser1Test.html">AddUser1Test</a></td></tr>
-<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser2Test.html">AddUser2Test</a></td></tr>
-<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser3Test.html">AddUser3Test</a></td></tr>
+<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUserTest.html">AddUserTest</a></td></tr>
+<tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUserPasswordTest.html">EditUserPasswordTest</a></td></tr>
+<tr><td><a href="SignOutTest.html">SignOutTest</a></td></tr>
+<tr><td><a href="User_SignInTest.html">User_SignInTest</a></td></tr>
+<tr><td><a href="SignOutTest.html">SignOutTest</a></td></tr>
+<tr><td><a href="SignInTest.html">SignInTest</a></td></tr>
 <tr><td><a href="../../../../../../portalweb/portal/controlpanel/users/user/adduser/TearDownUserTest.html">TearDownUserTest</a></td></tr>
 </tbody></table>
 </body>
