+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/signin/SignInTests.java
@@ -12,9 +12,12 @@
  * details.
  */
 
-package com.liferay.portalweb.portal.controlpanel.users.user.adduserpassword;
+package com.liferay.portalweb.portal.controlpanel.users.user.signin;
 
 import com.liferay.portalweb.portal.BaseTestSuite;
+import com.liferay.portalweb.portal.controlpanel.users.user.adduser.AddUserTest;
+import com.liferay.portalweb.portal.controlpanel.users.user.adduser.TearDownUserTest;
+import com.liferay.portalweb.portal.controlpanel.users.user.edituserpassword.EditUserPasswordTest;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -22,11 +25,15 @@ import junit.framework.TestSuite;
 /**
  * @author Brian Wing Shun Chan
  */
-public class AddUserPasswordTests extends BaseTestSuite {
+public class SignInTests extends BaseTestSuite {
 	public static Test suite() {
 		TestSuite testSuite = new TestSuite();
 		testSuite.addTestSuite(AddUserTest.class);
-		testSuite.addTestSuite(AddUserPasswordTest.class);
+		testSuite.addTestSuite(EditUserPasswordTest.class);
+		testSuite.addTestSuite(SignOutTest.class);
+		testSuite.addTestSuite(User_SignInTest.class);
+		testSuite.addTestSuite(SignOutTest.class);
+		testSuite.addTestSuite(SignInTest.class);
 		testSuite.addTestSuite(TearDownUserTest.class);
 
 		return testSuite;
