+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/signin/SignOutTest.html
@@ -1,0 +1,33 @@
+<html>
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+<title>SignOutTest</title>
+</head>
+<body>
+<table cellpadding="1" cellspacing="1" border="1">
+<thead>
+<tr><td rowspan="1" colspan="3">SignOutTest</td></tr>
+</thead><tbody>
+<tr>
+	<td>open</td>
+	<td>/web/guest/home/</td>
+	<td></td>
+</tr>
+<tr>
+	<td>waitForVisible</td>
+	<td>link=Sign Out</td>
+	<td>Sign Out</td>
+</tr>
+<tr>
+	<td>clickAtAndWait</td>
+	<td>link=Sign Out</td>
+	<td>Sign Out</td>
+</tr>
+<tr>
+	<td>waitForVisible</td>
+	<td>//input[@id='_58_login']</td>
+	<td>Email Address</td>
+</tr>
+</tbody></table>
+</body>
+</html>
