+++ portal-web/test/com/liferay/portalweb/portal/controlpanel/users/user/signin/User3_SignInTest.html
@@ -1,0 +1,48 @@
+<html>
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+<title>User3_SignInTest</title>
+</head>
+<body>
+<table cellpadding="1" cellspacing="1" border="1">
+<thead>
+<tr><td rowspan="1" colspan="3">User3_SignInTest</td></tr>
+</thead><tbody>
+<tr>
+	<td>open</td>
+	<td>/web/guest/home/</td>
+	<td></td>
+</tr>
+<tr>
+	<td>waitForVisible</td>
+	<td>link=Welcome</td>
+	<td>Welcome</td>
+</tr>
+<tr>
+	<td>clickAtAndWait</td>
+	<td>link=Welcome</td>
+	<td>Welcome</td>
+</tr>
+<tr>
+	<td>type</td>
+	<td>//input[@id='_58_login']</td>
+	<td>userea3@liferay.com</td>
+</tr>
+<tr>
+	<td>type</td>
+	<td>//input[@id='_58_password']</td>
+	<td>test</td>
+</tr>
+<tr>
+	<td>clickAt</td>
+	<td>//input[@id='_58_rememberMeCheckbox']</td>
+	<td>Remember Me</td>
+</tr>
+<tr>
+	<td>clickAtAndWait</td>
+	<td>//input[@value='Sign In']</td>
+	<td>Sign In</td>
+</tr>
+</tbody></table>
+</body>
+</html>
