+++ portal-web/test/com/liferay/portalweb/portal/permissions/usecase/permissionsgroupcompanydemo/PermissionsGroupCompanyDemoTests.html
@@ -11,9 +11,9 @@
 <tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPageBlogsTest.html">AddPageBlogsTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/blogs/portlet/addportletblogs/AddPortletBlogsTest.html">AddPortletBlogsTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser1Test.html">AddUser1Test</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUser1PasswordTest.html">EditUser1PasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUser1PasswordTest.html">EditUser1PasswordTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser2Test.html">AddUser2Test</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUser2PasswordTest.html">EditUser2PasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUser2PasswordTest.html">EditUser2PasswordTest</a></td></tr>
 <tr><td><a href="AssignMembersUser1DefaultSiteTest.html">AssignMembersUser1DefaultSiteTest</a></td></tr>
 <tr><td><a href="AddRegularRoleBlogsEditorTest.html">AddRegularRoleBlogsEditorTest</a></td></tr>
 <tr><td><a href="DefinePermissionsRoleSiteMemberTest.html">DefinePermissionsRoleSiteMemberTest</a></td></tr>
