+++ portal-web/test/com/liferay/portalweb/portal/permissions/usecase/permissionsindividualscopedemo/PermissionsIndividualScopeDemoTests.html
@@ -9,9 +9,9 @@
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
 <tr><td><b>Test Suite</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser1Test.html">AddUser1Test</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUser1PasswordTest.html">EditUser1PasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUser1PasswordTest.html">EditUser1PasswordTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser2Test.html">AddUser2Test</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUser2PasswordTest.html">EditUser2PasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUser2PasswordTest.html">EditUser2PasswordTest</a></td></tr>
 <tr><td><a href="EditUser2SitesLiferayTest.html">EditUser2SitesLiferayTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentlibrary/portlet/addportlet/AddPageDMTest.html">AddPageDMTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentlibrary/portlet/addportlet/AddPortletDMTest.html">AddPortletDMTest</a></td></tr>
