+++ portal-web/test/com/liferay/portalweb/portal/permissions/usecase/permissionsteamdemo/PermissionsTeamDemoTests.html
@@ -11,9 +11,9 @@
 <tr><td><a href="AddSiteTest.html">AddSiteTest</a></td></tr>
 <tr><td><a href="AddPageSiteTest.html">AddPageSiteTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser1Test.html">AddUser1Test</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUser1PasswordTest.html">EditUser1PasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUser1PasswordTest.html">EditUser1PasswordTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUser2Test.html">AddUser2Test</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUser2PasswordTest.html">EditUser2PasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUser2PasswordTest.html">EditUser2PasswordTest</a></td></tr>
 <tr><td><a href="AssignMembersUser1SiteTest.html">AssignMembersUser1SiteTest</a></td></tr>
 <tr><td><a href="AssignMembersUser2SiteTest.html">AssignMembersUser2SiteTest</a></td></tr>
 <tr><td><a href="AssignRoleSiteAdministratorUser1Test.html">AssignRoleSiteAdministratorUser1Test</a></td></tr>
