+++ portal-web/test/com/liferay/portalweb/portal/permissions/usecase/permissionsuserpersonalsitedemo/PermissionsUserPersonalSiteDemoTests.html
@@ -19,7 +19,7 @@
 <tr><td><a href="DeletePermissionsUserBlogsPermissionsUPSTest.html">DeletePermissionsUserBlogsPermissionsUPSTest</a></td></tr>
 <tr><td><a href="DeletePermissionsUserBlogsSubscribeUPSTest.html">DeletePermissionsUserBlogsSubscribeUPSTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/AddUserTest.html">AddUserTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/adduser/EditUserPasswordTest.html">EditUserPasswordTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portal/controlpanel/users/user/edituserpassword/EditUserPasswordTest.html">EditUserPasswordTest</a></td></tr>
 <tr><td><a href="SignOutTest.html">SignOutTest</a></td></tr>
 <tr><td><a href="User_SignInTest.html">User_SignInTest</a></td></tr>
 <tr><td><a href="User_ViewOffPermissionsBlogsUSPTest.html">User_ViewOffPermissionsBlogsUSPTest</a></td></tr>
