+++ portal-web/test/com/liferay/portalweb/portal/permissions/usecase/permissionsuserpersonalsitedemo/PermissionsUserPersonalSiteDemoTests.java
@@ -16,8 +16,8 @@ package com.liferay.portalweb.portal.permissions.usecase.permissionsuserpersonal
 
 import com.liferay.portalweb.portal.BaseTestSuite;
 import com.liferay.portalweb.portal.controlpanel.users.user.adduser.AddUserTest;
-import com.liferay.portalweb.portal.controlpanel.users.user.adduser.EditUserPasswordTest;
 import com.liferay.portalweb.portal.controlpanel.users.user.adduser.TearDownUserTest;
+import com.liferay.portalweb.portal.controlpanel.users.user.edituserpassword.EditUserPasswordTest;
 import com.liferay.portalweb.portal.util.TearDownPageTest;
 import com.liferay.portalweb.portlet.blogs.portlet.addportletblogs.AddPageBlogsTest;
 import com.liferay.portalweb.portlet.blogs.portlet.addportletblogs.AddPortletBlogsTest;
